Changelog: Se hicieron cambios en el id del growl para el formulario de creación de rutina. Se implementó en la misma vista nuevaRutina.xhtml el formulario para agregar días a esa rutina. se usa un nuevo growl con un id diferente al growl del formulario de rutina. Se implementaron botones para limpiar campos tanto para la rutina como para el día. Se removió el componente que mostraba mensajes antes del formulario de rutina, ahora todos los mensajes se muestran en un growl con una vida más larga. Se pueden agregar días a una rutina.

+++ src/java/co/edu/unicauca/gymadmdoc/sessionBeans/RutinasDeUsuarioBean.java
@@ -7,6 +7,7 @@ package co.edu.unicauca.gymadmdoc.sessionBeans;
 
 import co.edu.unicauca.gymadmdoc.entities.MruRutina;
 import java.util.List;
+import javax.ejb.EJB;
 import javax.ejb.Stateless;
 import javax.persistence.EntityManager;
 import javax.persistence.PersistenceContext;
@@ -23,8 +24,9 @@ public class RutinasDeUsuarioBean {
     // "Insert Code > Add Business Method")
     @PersistenceContext(unitName="Gym_Adm_DocPU")
     private EntityManager em;
-    
-    public List<MruRutina> consultarRutinasPorUsuario(float usuId){
+
+    public List<MruRutina> consultarRutinasPorUsuario(Long usuId){
+       
         TypedQuery<MruRutina> consulta = (TypedQuery<MruRutina>) getEm().createNamedQuery("MruRutina.findByUsuId");
         consulta.setParameter("usuIdentificacion", usuId);
         List<MruRutina> resultList = consulta.getResultList();
